+++ packages/nimble-components/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log - @ni/nimble-components
 
-This log was last generated on Fri, 08 Sep 2023 23:00:33 GMT and should not be manually modified.
+This log was last generated on Tue, 12 Sep 2023 17:22:19 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 20.2.8
+
+Tue, 12 Sep 2023 17:22:19 GMT
+
+### Patches
+
+- Add locale support to number-text column ([ni/nimble@b49755b](https://github.com/ni/nimble/commit/b49755bcb4948910c5b896b7c11abb64c6aca2cf))
 
 ## 20.2.7